fix(demos): cast gaia camera near/far to float (JSON-serialization robustness)
Double-check caught that the gaia CameraConfig near/far were left as numpy
float64 (from np.tan/np.max) while position/target were float()-cast. It
serialized fine, but np.float64 in a JSON-serialized viewer config is fragile
and inconsistent — cast both to plain float.

+++ packages/luxar/src/luxar/demos/demo_gaia_milky_way_3m.py
@@ -261,8 +261,8 @@ def load_and_convert_gaia_data(data_zarr_path: Path, output_path: Path) -> int:
             target=(float(center[0]), float(center[1]), float(center[2])),
             up=(0.0, 1.0, 0.0),
             fov=fov_deg,
-            near=max(0.5, cam_dist * 0.005),
-            far=cam_dist * 20.0 + extent * 10.0,
+            near=float(max(0.5, cam_dist * 0.005)),
+            far=float(cam_dist * 20.0 + extent * 10.0),
         )
 
         with LuxarZarrCompiler(output_path) as compiler:
